random bugfixes
not describing as significant changes are coming soon

+++ includes/Color.php
@@ -685,7 +685,7 @@ class Color {
 		return array_map("hexdec", $exploded);
 	}
 
-	public static function lightenHex(string $hex, int $steps) : string {
+	public static function lightenHex(string $hex, float $steps) : string {
 		if (strlen($hex) == 3) {
 			$hex = $hex[0].$hex[0].$hex[1].$hex[1].$hex[2].$hex[2];
 		}
